Updates v1

+++ OTFood-ejb/src/java/ejb/session/stateless/CreditCardEntitySessionBean.java
@@ -7,13 +7,14 @@ package ejb.session.stateless;
 
 import entity.CreditCardEntity;
 import entity.OTUserEntity;
+import entity.SaleTransactionEntity;
 import java.util.List;
+import java.util.Objects;
 import java.util.Set;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 import javax.ejb.EJB;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
+import javax.persistence.NoResultException;
 import javax.persistence.PersistenceContext;
 import javax.persistence.PersistenceException;
 import javax.persistence.Query;
@@ -26,7 +27,6 @@ import util.exception.CreditCardExistException;
 import util.exception.CreditCardNotFoundException;
 import util.exception.InputDataValidationException;
 import util.exception.UnknownPersistenceException;
-import util.exception.UserExistException;
 import util.exception.UserNotFoundException;
 
 /**
@@ -51,27 +51,34 @@ public class CreditCardEntitySessionBean implements CreditCardEntitySessionBeanL
     }
 
     @Override
-    public CreditCardEntity createNewCreditCardForUser(CreditCardEntity card, Long userId) throws InputDataValidationException, UnknownPersistenceException, CreditCardExistException, CardCreationException, UserNotFoundException {
+    public CreditCardEntity createNewCreditCardForUser(CreditCardEntity card, Long userId) throws InputDataValidationException, UnknownPersistenceException, CreditCardExistException, UserNotFoundException {
         Set<ConstraintViolation<CreditCardEntity>> constraintViolations = validator.validate(card);
-
         if (constraintViolations.isEmpty()) {
+            OTUserEntity user = oTUserEntitySessionBean.retrieveUserById(userId);
+            String cardNumber = card.getCardNumber();
+            Query query = em.createQuery("SELECT c FROM CreditCardEntity c WHERE c.cardNumber = :ccNum AND c.isRemoved = true");
+            query.setParameter("ccNum", cardNumber);
             try {
-                if (userId != null) {
-                    try {
-                        OTUserEntity user = oTUserEntitySessionBean.retrieveUserById(userId);
+                CreditCardEntity creditCard = (CreditCardEntity) query.getSingleResult();
+                creditCard.setIsRemoved(false);
+                creditCard.setCardName(card.getCardName());
+                creditCard.setExpiryDate(card.getExpiryDate());
+                creditCard.setType(card.getType());
 
-                        em.persist(card);
-                        user.getCreditCard().add(card);
-                        card.setUser(user);
-
-                        em.flush();
-                        return card;
-                    } catch (UserNotFoundException ex) {
-                        throw new UserNotFoundException(ex.getMessage());
-                    }
-                } else {
-                    throw new CardCreationException("User ID must be provided!");
-                }
+                creditCard.getUser().getCreditCard().remove(creditCard); // can remove because creditCardId has already been set
+                creditCard.setUser(user);
+                user.getCreditCard().add(creditCard);
+                em.flush();
+                return creditCard;
+            } catch (NoResultException ex) {
+                // will continue to below in this case
+            }
+            try {
+                em.persist(card);
+                user.getCreditCard().add(card);
+                card.setUser(user);
+                em.flush();
+                return card;
             } catch (PersistenceException ex) {
                 if (ex.getCause() != null && ex.getCause().getClass().getName().equals("org.eclipse.persistence.exceptions.DatabaseException")) {
                     if (ex.getCause().getCause() != null && ex.getCause().getCause().getClass().getName().equals("java.sql.SQLIntegrityConstraintViolationException")) {
@@ -92,7 +99,7 @@ public class CreditCardEntitySessionBean implements CreditCardEntitySessionBeanL
     public CreditCardEntity retrieveCardById(Long id) throws CreditCardNotFoundException {
         CreditCardEntity card = em.find(CreditCardEntity.class, id);
 
-        if (card != null) {
+        if (card != null && card.isIsRemoved() == false) {
             return card;
         } else {
             throw new CreditCardNotFoundException("Credit Card ID " + id + " does not exist!");
@@ -101,7 +108,7 @@ public class CreditCardEntitySessionBean implements CreditCardEntitySessionBeanL
 
     @Override
     public List<CreditCardEntity> retrieveAllCardByUser(Long userId) {
-        Query query = em.createQuery("SELECT c FROM CreditCardEntity c WHERE c.user.UserId = :id");
+        Query query = em.createQuery("SELECT c FROM CreditCardEntity c WHERE c.user.UserId = :id AND c.isRemoved = false");
         query.setParameter("id", userId);
 
         return query.getResultList();
@@ -109,7 +116,7 @@ public class CreditCardEntitySessionBean implements CreditCardEntitySessionBeanL
 
     @Override
     public CreditCardEntity retrieveCardByCardNumber(String cardNumber) {
-        Query query = em.createQuery("SELECT c FROM CreditCardEntity c WHERE c.cardNumber = :num");
+        Query query = em.createQuery("SELECT c FROM CreditCardEntity c WHERE c.cardNumber = :num AND c.isRemoved = false");
         query.setParameter("num", cardNumber);
 
         return (CreditCardEntity) query.getSingleResult();
@@ -120,12 +127,21 @@ public class CreditCardEntitySessionBean implements CreditCardEntitySessionBeanL
         try {
             CreditCardEntity card = retrieveCardById(id);
 
-            if (card.getUser() != null) {
+            boolean toRemove = true;
+            List<SaleTransactionEntity> saleTransactions = card.getUser().getSaleTransaction();
+            for (SaleTransactionEntity st : saleTransactions) {
+                if (Objects.equals(st.getCreditCardEntity().getCreditCardId(), id)) {
+                    toRemove = false;
+                    break;
+                }
+            }
+            if (toRemove) {
                 OTUserEntity user = card.getUser();
                 user.getCreditCard().remove(card);
-            }
-
-            em.remove(card);
+                em.remove(card);
+            } else {
+                card.setIsRemoved(true);
+            }              
         } catch (CreditCardNotFoundException ex) {
             throw new CreditCardNotFoundException(ex.getMessage());
         }
@@ -133,11 +149,9 @@ public class CreditCardEntitySessionBean implements CreditCardEntitySessionBeanL
 
     private String prepareInputDataValidationErrorsMessage(Set<ConstraintViolation<CreditCardEntity>> constraintViolations) {
         String msg = "Input data validation error!:";
-
         for (ConstraintViolation constraintViolation : constraintViolations) {
             msg += "\n\t" + constraintViolation.getPropertyPath() + " - " + constraintViolation.getInvalidValue() + "; " + constraintViolation.getMessage();
         }
-
         return msg;
     }
 }
